MINC 2016-02-24 (8632513e)
Run the UpdateFromUpstream.sh script to extract upstream MINC
using the following shell commands.

$ git archive --prefix=upstream-minc/ 8632513e -- 
./ChangeLog
./volume_io
./libsrc
./libsrc/minc_compat.h
./libsrc/minc_config.h
./libsrc/minc_config.c
./libsrc/minc_format_convert.c
./libsrc/value_conversion.c
./libsrc/hdf_convenience.c
./libsrc/read_file_names.c
./libsrc/minc_simple.c
./libsrc/time_stamp.h
./libsrc/hdf_convenience.h
./libsrc/type_limits.h
./libsrc/read_file_names.h
./libsrc/minc_varlists.h
./libsrc/restructure.h
./libsrc/nd_loop.h
./libsrc/restructure.c
./libsrc/minc_simple.h
./libsrc/strdup.c
./libsrc/minc_useful.h
./libsrc/minc_convenience.c
./libsrc/minc_basic.h
./libsrc/minc_compat.c
./libsrc/time_stamp.c
./libsrc/voxel_loop.h
./libsrc/minc_routines.h
./libsrc/minc_private.h
./libsrc/netcdf_convenience.c
./libsrc/voxel_loop.c
./libsrc/dim_conversion.c
./libsrc/minc_format_convert.h
./libsrc/nd_loop.c
./libsrc/ParseArgv.h
./libsrc/minc_error.c
./libsrc/minc_error.h
./libsrc/minc_structures.h
./libsrc/minc.h
./libsrc/image_conversion.c
./libsrc/ParseArgv.c
./COPYING
./UseLIBMINC.cmake.in
./NEWS
./AUTHORS
./libsrc2/m2util.c
./libsrc2/minc2_defs.h
./libsrc2/minc2_api.h
./libsrc2/minc2_error.h
./libsrc2/grpattr.c
./libsrc2/hyper.c
./libsrc2/minc_compat2.h
./libsrc2/minc2_error.c
./libsrc2/record.c
./libsrc2/volume.c
./libsrc2/datatype.c
./libsrc2/volprops.c
./libsrc2/valid.c
./libsrc2/convert.c
./libsrc2/free.c
./libsrc2/dimension.c
./libsrc2/label.c
./libsrc2/minc2.h
./libsrc2/minc2_private.h
./libsrc2/slice.c
./libsrc2/minc2_structs.h
./CMakeLists.txt
./nifti
./LIBMINCConfig.cmake.in
./README.release
./INSTALL
./config.h.cmake
./README
 | tar x
$ git shortlog --no-merges --abbrev=8 --format='%h %s' dcb93a5f..8632513e

Gabriel A. Devenyi (1):
      9d9c184c Fix check for round and copysign

Robert D. Vincent (13):
      85db0e34 Use external NIfTI-1 library from sourceforge by default.
      ab99a306 Set LIBMINC_NIFTI_SUPPORT if appropriate.
      3847385d Use -DBL_MAX for maximum search.
      36d6d5c5 Correct comment for miget_attr_length()
      783bca38 Update doc.
      726c0af2 Fix lack of explicit NIFTI dependencies.
      47beebd5 Fix handling of dimension widths for irregular dimensions.
      e515b43a Hopefully fix test failures with ASan.
      2c574209 Fix illegal memory read.
      969631e8 One last issue, hopefully, before ASan is happy.
      4a4539f7 Fix stupid error.
      a1839e0d Fix for proper handling of RGB volumes.
      7d6ac4e1 Attempt to read DT_RGB24 NIfTI images.

Sean McBride (3):
      e778c447 Fixed several off-by-1 bugs, found by ASan
      458997c6 Fixed some -Wdouble-promotion warnings with casts
      68d3cc2b Fixed -Wdouble-promotion warnings changing FLT_MAX to DBL_MAX

Vladimir S. FONOV (4):
      ba5f123e Updated NIFTI building script
      7f71fc85 Made dependency only if this is a standalone libminc compilation
      e56afda6 Made changes to specify NIFTI in LIBMINC configuration files
      8632513e Removed obsolete settings for RPATH, seem to be working as needed without it

+++ libsrc2/hyper.c
@@ -340,9 +340,9 @@ cleanup:
       for(_j=0;_j<image_slice_length;_j++)\
       {\
         double _temp;\
-        _temp=*_buffer;\
-        _temp= _temp*_scale  + _offset ; \
-        *_buffer =(type_in)(_temp); \
+        _temp=(double)*_buffer;\
+        _temp=_temp*_scale  + _offset ; \
+        *_buffer=(type_in)(_temp); \
         _buffer++;\
       }\
     }\
@@ -801,7 +801,7 @@ cleanup:
     hsize_t _i,_j;\
     double voxel_offset=voxel_min;\
     double voxel_range=voxel_max-voxel_min;\
-    double norm_offset=norm_min;\
+    double norm_offset=(double)norm_min;\
     double norm_range=(double)norm_max-(double)norm_min;\
     double data_offset=data_min;\
     double data_range=(double)data_max-(double)data_min;\
